# docs/api/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_anychat_server_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  internal_gateway_handler.ackReadTriggersRequest:
    properties:
      events:
        items:
          $ref: '#/definitions/internal_gateway_handler.readTriggerEvent'
        minItems: 1
        type: array
    required:
    - events
    type: object
...
  internal_gateway_handler.readTriggerEvent:
    properties:
      client_at:
        type: integer
      idempotency_key:
        type: string
      message_id:
        type: string
    required:
    - message_id
    type: object
  internal_gateway_handler.recallMessageRequest:
    properties:
      message_id:
        type: string
    required:
    - message_id
    type: object
  internal_gateway_handler.sendMessageRequest:
    properties:
      at_users:
        items:
          type: string
        type: array
      content:
        type: string
      content_type:
        enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        - 7
        type: integer
      conversation_id:
        type: string
      local_id:
        type: string
      reply_to:
        type: string
    required:
    - content
    - content_type
    - conversation_id
    - local_id
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@anychat.example.com
    name: AnyChat API Support
    url: https://github.com/yzhgit/anychat-server
  description: |-
    AnyChat instant messaging system gateway API service, providing HTTP interfaces for user authentication, user management, and other functions.
    All endpoints requiring authentication must include Authorization: Bearer <token> in the Header.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: AnyChat Gateway API
  version: "1.0"
paths:
...
  /conversations/{conversationId}/messages/after:
    get:
      consumes:
      - application/json
      description: Query messages after an anchor_message_id in a conversation
      parameters:
      - description: conversation ID
        in: path
        name: conversationId
        required: true
        type: string
      - description: anchor message ID
        in: query
        name: anchor_message_id
        required: true
        type: string
      - description: limit (default 20, max 100)
        format: int32
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: parameter error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: get messages after anchor
      tags:
      - message
  /conversations/{conversationId}/messages/around-anchor:
    get:
      consumes:
      - application/json
      description: Query messages before and after anchor_message_id in one request
      parameters:
      - description: conversation ID
        in: path
        name: conversationId
        required: true
        type: string
      - description: anchor message ID
        in: query
        name: anchor_message_id
        required: true
        type: string
      - description: messages before anchor (default 20, max 100)
        format: int32
        in: query
        name: before
        type: integer
      - description: messages after anchor (default 20, max 100)
        format: int32
        in: query
        name: after
        type: integer
      - description: include anchor message (default true)
        in: query
        name: include_anchor
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: parameter error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: get messages around anchor
      tags:
      - message
  /conversations/{conversationId}/messages/before:
    get:
      consumes:
      - application/json
      description: Query messages before an anchor_message_id in a conversation
      parameters:
      - description: conversation ID
        in: path
        name: conversationId
        required: true
        type: string
      - description: anchor message ID
        in: query
        name: anchor_message_id
        required: true
        type: string
      - description: limit (default 20, max 100)
        format: int32
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: parameter error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: get messages before anchor
      tags:
      - message
  /conversations/{conversationId}/messages/first-unread-anchor:
    get:
      consumes:
      - application/json
      description: Query first unread message anchor, optionally returning context
        windows
      parameters:
      - description: conversation ID
        in: path
        name: conversationId
        required: true
        type: string
      - description: whether to include before/after context
        in: query
        name: with_context
        type: boolean
      - description: messages before anchor when with_context=true (default 20, max
          100)
        format: int32
        in: query
        name: before
        type: integer
      - description: messages after anchor when with_context=true (default 20, max
          100)
        format: int32
        in: query
        name: after
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: parameter error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: get first unread anchor
      tags:
      - message
...
  /messages:
    post:
      consumes:
      - application/json
      description: Send conversation message via HTTP (supports idempotent local_id)
      parameters:
      - description: message content
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_gateway_handler.sendMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: parameter error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: send message
      tags:
      - message
  /messages/{messageId}:
    delete:
      consumes:
      - application/json
      description: Delete specified message, can only delete own messages
      parameters:
      - description: message ID
        in: path
        name: messageId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "403":
          description: no permission
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "404":
          description: message not found
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: delete message
      tags:
      - message
    get:
      consumes:
      - application/json
      description: Get single message by message ID
      parameters:
      - description: message ID
        in: path
        name: messageId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: parameter error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "404":
          description: message not found
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: get message detail
      tags:
      - message
  /messages/read-triggers:
    post:
      consumes:
      - application/json
      description: Client batch reports message read trigger events, server starts
        burn timer accordingly
      parameters:
      - description: read trigger events
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_gateway_handler.ackReadTriggersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: parameter error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: burn after reading read trigger acknowledgment
      tags:
      - message
  /messages/recall:
    post:
      consumes:
      - application/json
      description: Recall specified message, can only recall own messages within 2
        minutes
      parameters:
      - description: message ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_gateway_handler.recallMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "400":
          description: parameter error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "403":
          description: no permission or timeout
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "404":
          description: message not found
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: recall message
      tags:
      - message
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"